requirements.txt
# C盘大文件迁移工具依赖包

# GUI界面
PyQt5>=5.15.0

# 配置文件
pyyaml>=6.0

# 其他依赖
psutil>=5.9.0  # 用于进程管理
pathlib>=1.0.1  # 用于路径处理
tqdm>=4.64.0  # 用于进度显示